+++ src/styleguide/setup/routes.ts
@@ -19,6 +19,7 @@ export default [
     redirect: '/sg/sg-test-page-readme',
     meta: {
       title: 'Styleguide',
+      sortOrder: 1,
     },
     children: [
       styleguideTestPages.readme,
@@ -48,6 +49,7 @@ export default [
     component: styleguideRouterConfig.routeChildrenComponentWrapper,
     meta: {
       title: 'Components',
+      sortOrder: 3,
     },
     children: [
       {
@@ -56,6 +58,7 @@ export default [
         component: () => import('../demo-pages/components/r-navigation.vue'),
         meta: {
           title: 'Navigation',
+          sortOrder: 2,
         },
       },
       {
@@ -64,6 +67,7 @@ export default [
         component: () => import('../demo-pages/components/r-vas-modal.vue'),
         meta: {
           title: 'Modal',
+          sortOrder: 1,
         },
       },
       {
@@ -90,6 +94,7 @@ export default [
     component: styleguideRouterConfig.routeChildrenComponentWrapper,
     meta: {
       title: 'Elements',
+      sortOrder: 2,
     },
     children: [
       {